+++ app/src/main/java/com/sdhong/pokemonapp/view/MainActivity.kt
@@ -7,7 +7,6 @@ import androidx.appcompat.content.res.AppCompatResources
 import androidx.core.view.ViewCompat
 import androidx.core.view.WindowInsetsCompat
 import com.google.android.material.tabs.TabLayoutMediator
-import com.sdhong.pokemonapp.ViewPagerAdapter
 import com.sdhong.pokemonapp.databinding.ActivityMainBinding
 import com.sdhong.pokemonapp.local.MainTab
